# public/swagger-ui/swagger.json
{
    "swagger": "2.0",
    "info": {  
        "description": "Project to manage information about fields, tractors and processed fields",
        "version": "1.0.0",
        "title": "Manage Agriculture Field",
        "contact": {  
            "email": "deepakms001@yahoo.com"
        }
    },
    "host": "localhost:8000",
    "basePath": "",
...
    "schemes": [  
        "http"
    ],
    "paths": {  
...
        "/tractors": {  
            "get": {  
                "tags": [  
                    "Tractors"
                ],
                "summary": "Get all tractors",
                "description": "Get all tractor list",
                "operationId": "/allTractors",
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "query",
                        "description": "Token obtained from authentication API ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {  
                    "200": {  
                        "description": "successful operation",
                        "schema": {  
                            "type": "array",
                            "items": {  
                                "$ref": "#/definitions/Pet"
                            }
                        }
                    },
                    "400": {  
                        "description": "Provided token is expired",
                        "schema": {  
                            "type": "object",
                            "items": {
                                "status": "error",
                                "message": "Provided token is expired"
                            }
                        }
                    }
                }
            },
            "post": {  
                "tags": [  
                    "Tractors"
                ],
                "summary": "Add a Tractor",
                "description": "",
                "operationId": "addTractor",
                "consumes": [  
                    "application/x-www-form-urlencoded"
                ],
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "name",
                        "in": "formData",
                        "description": "Name of the Tractor",
                        "required": true,
                        "type": "string"
                    }, {  
                        "name": "reg_number",
                        "in": "formData",
                        "description": "UID of the Tractor",
                        "required": true,
                        "type": "string"
                    },
                    {  
                        "name": "token",
                        "in": "formData",
                        "description": "Token optained from auth API",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {  
                    "405": {  
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/tractors/{tractorId}": {
            "delete": {  
                "tags": [  
                    "Tractors"
                ],
                "summary": "Delete a tractor",
                "description": "",
                "operationId": "deleteTractor",
                "produces": [  
                    "application/json"
                ],
                "parameters": [  
                    {  
                        "name": "token",
                        "in": "query",
                        "description": "Token obtained from auth API",
                        "required": true,
                        "type": "string"
                    },
                    {  
                        "name": "tractorId",
                        "in": "path",
                        "description": "ID of tractor to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {  
                    "400": {  
                        "description": "Invalid ID supplied"
                    },
                    "404": {  
                        "description": "Pet not found"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {  

        "token": {  
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    },
    "definitions": {  
...
        "Category": {  
            "type": "object",
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "name": {  
                    "type": "string"
                }
            },
            "xml": {  
                "name": "Category"
            }
        },
        "Tag": {  
            "type": "object",
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "name": {  
                    "type": "string"
                }
            },
            "xml": {  
                "name": "Tag"
            }
        },
        "Pet": {  
            "type": "object",
            "required": [  
                "name",
                "photoUrls"
            ],
            "properties": {  
                "id": {  
                    "type": "integer",
                    "format": "int64"
                },
                "category": {  
                    "$ref": "#/definitions/Category"
                },
                "name": {  
                    "type": "string",
                    "example": "doggie"
                },
                "photoUrls": {  
                    "type": "array",
                    "xml": {  
                        "name": "photoUrl",
                        "wrapped": true
                    },
                    "items": {  
                        "type": "string"
                    }
                },
                "tags": {  
                    "type": "array",
                    "xml": {  
                        "name": "tag",
                        "wrapped": true
                    },
                    "items": {  
                        "$ref": "#/definitions/Tag"
                    }
                },
                "status": {  
                    "type": "string",
                    "description": "pet status in the store",
                    "enum": [  
                        "available",
                        "pending",
                        "sold"
                    ]
                }
            },
            "xml": {  
                "name": "Pet"
            }
        },
...
    },
...
}